Properly split tests that require Scala 3
